feat(modularization): extract memory ingest/recall into clawql-memory (phase 7)
Move runMemoryIngest, runMemoryRecall, and enterprise citations into clawql-memory;
keep MCP handlers in src/ with mcp-tool-log shaping only at the transport layer.
