+++ packages/clawql-memory/tsup.config.ts
@@ -9,6 +9,9 @@ export default defineConfig({
     "vault/provider-index": "src/vault/provider-index.ts",
     "ingest/slug": "src/ingest/slug.ts",
     "ingest/hashes": "src/ingest/hashes.ts",
+    "ingest/enterprise-citations": "src/ingest/enterprise-citations.ts",
+    "ingest/ingest": "src/ingest/ingest.ts",
+    "recall/recall": "src/recall/recall.ts",
     "embedding/embedding": "src/embedding/embedding.ts",
     "db/artifacts": "src/db/artifacts.ts",
     "db/postgres-migrations": "src/db/postgres-migrations.ts",
